+++ ElvUI_MerathilisUI/Core/Changelog/7.06.lua
@@ -1,8 +1,14 @@
 local MER = unpack(ElvUI_MerathilisUI)
 
 MER.Changelog[706] = {
-	RELEASE_DATE = "TBD",
-	FIXES = {},
-	NEW = {},
-	IMPROVEMENTS = {},
+	RELEASE_DATE = "26.01.2026",
+	FIXES = {
+		"[Install]: Remove cutaway settings for UnitFrames.",
+	},
+	NEW = {
+		"[DataTexts]: Added a mew Durability_iLevel DataText from mMediaTag",
+	},
+	IMPROVEMENTS = {
+		"[Tag]: Protect the [name:gradient] tag a bit because of secret values.",
+	},
 }